refactor: convert namespace.php function files to classes
Addresses PR feedback: replace the theme's eight namespace.php files
with PSR-4 classes (Actions, Activation, Admin\Catalog, Customizer,
Filters, Shortcodes) plus a single src/helpers.php.

Helpers remain functions because partials and templates call them at
render time (including the contact-form handler and the default_menu
fallback callable). The Aldine\Admin\BLOG_OPTION constant stays a
namespace constant in helpers.php because pressbooks core detects it
via defined(), as does Aldine\Customizer\MAX_FEATURED_BOOKS. The
empty Tags file is removed. All hooks in functions.php and the
TinyMCE button registrations inside Actions now use
[ClassName::class, 'method'] callables, and test-filters.php calls
Filters::registerQueryVars.
