+++ functions.php
@@ -23,51 +23,51 @@ if (! file_exists($composer)) {
 }
 require_once $composer;
 
-add_action('after_switch_theme', '\Aldine\Activation\create_default_content', 10);
-add_action('after_switch_theme', '\Aldine\Activation\create_menus', 11);
-add_action('after_switch_theme', '\Aldine\Activation\assign_menus', 12);
-add_action('admin_bar_init', '\Aldine\Actions\remove_admin_bar_callback');
-add_action('after_setup_theme', '\Aldine\Actions\setup');
-add_action('after_setup_theme', '\Aldine\Actions\content_width', 0);
-add_action('wp_head', '\Aldine\Actions\output_custom_colors');
-add_action('init', '\Aldine\Actions\add_editor_styles');
-add_action('admin_init', '\Aldine\Actions\hide_catalog_content_editor');
+add_action('after_switch_theme', [\Aldine\Activation::class, 'createDefaultContent'], 10);
+add_action('after_switch_theme', [\Aldine\Activation::class, 'createMenus'], 11);
+add_action('after_switch_theme', [\Aldine\Activation::class, 'assignMenus'], 12);
+add_action('admin_bar_init', [\Aldine\Actions::class, 'removeAdminBarCallback']);
+add_action('after_setup_theme', [\Aldine\Actions::class, 'setup']);
+add_action('after_setup_theme', [\Aldine\Actions::class, 'contentWidth'], 0);
+add_action('wp_head', [\Aldine\Actions::class, 'outputCustomColors']);
+add_action('init', [\Aldine\Actions::class, 'addEditorStyles']);
+add_action('admin_init', [\Aldine\Actions::class, 'hideCatalogContentEditor']);
 foreach ([ 'post.php', 'post-new.php' ] as $hook) {
-    add_action("admin_head-$hook", '\Aldine\Actions\tinymce_l18n');
+    add_action("admin_head-$hook", [\Aldine\Actions::class, 'tinymceL18N']);
 }
-add_filter('body_class', '\Aldine\Filters\body_classes');
-add_filter('excerpt_more', '\Aldine\Filters\excerpt_more');
-add_filter('query_vars', '\Aldine\Filters\register_query_vars');
-add_filter('wp_nav_menu_items', '\Aldine\Filters\adjust_menu', 10, 2);
+add_filter('body_class', [\Aldine\Filters::class, 'bodyClasses']);
+add_filter('excerpt_more', [\Aldine\Filters::class, 'excerptMore']);
+add_filter('query_vars', [\Aldine\Filters::class, 'registerQueryVars']);
+add_filter('wp_nav_menu_items', [\Aldine\Filters::class, 'adjustMenu'], 10, 2);
 add_filter('the_content', 'apply_shortcodes');
 add_filter('show_admin_bar', '__return_false');
-add_action('widgets_init', '\Aldine\Actions\widgets_init');
-add_action('widgets_init', '\Aldine\Actions\remove_widgets');
-add_action('wp_enqueue_scripts', '\Aldine\Actions\enqueue_assets');
-add_action('updated_option', '\Aldine\Actions\add_color_variants', 10, 3);
-add_action('customize_register', '\Aldine\Customizer\customize_register');
-add_action('customize_preview_init', '\Aldine\Customizer\customize_preview_js');
-add_action('customize_controls_enqueue_scripts', '\Aldine\Customizer\enqueue_color_contrast_validator');
-add_action('customize_controls_enqueue_scripts', '\Aldine\Customizer\featured_books_scripts');
-add_action('customize_controls_enqueue_scripts', '\Aldine\Customizer\enqueue_contact_form_tweaks');
-add_action('customize_controls_enqueue_scripts', '\Aldine\Customizer\enqueue_pb_a11y_in_customizer');
-add_action('customize_controls_enqueue_scripts', '\Aldine\Customizer\enqueue_catalog_search_control_assets');
-add_action('wp_ajax_pb_search_catalog_books', '\Aldine\Customizer\ajax_search_catalog_books');
+add_action('widgets_init', [\Aldine\Actions::class, 'widgetsInit']);
+add_action('widgets_init', [\Aldine\Actions::class, 'removeWidgets']);
+add_action('wp_enqueue_scripts', [\Aldine\Actions::class, 'enqueueAssets']);
+add_action('updated_option', [\Aldine\Actions::class, 'addColorVariants'], 10, 3);
+add_action('customize_register', [\Aldine\Customizer::class, 'customizeRegister']);
+add_action('customize_preview_init', [\Aldine\Customizer::class, 'customizePreviewJs']);
+add_action('customize_controls_enqueue_scripts', [\Aldine\Customizer::class, 'enqueueColorContrastValidator']);
+add_action('customize_controls_enqueue_scripts', [\Aldine\Customizer::class, 'featuredBooksScripts']);
+add_action('customize_controls_enqueue_scripts', [\Aldine\Customizer::class, 'enqueueContactFormTweaks']);
+add_action('customize_controls_enqueue_scripts', [\Aldine\Customizer::class, 'enqueuePbA11YInCustomizer']);
+add_action('customize_controls_enqueue_scripts', [\Aldine\Customizer::class, 'enqueueCatalogSearchControlAssets']);
+add_action('wp_ajax_pb_search_catalog_books', [\Aldine\Customizer::class, 'ajaxSearchCatalogBooks']);
 
 // Shortcodes.
-add_shortcode('aldine_page_section', '\Aldine\Shortcodes\page_section');
-add_shortcode('aldine_call_to_action', '\Aldine\Shortcodes\call_to_action');
-add_shortcode('latest-titles', '\Aldine\Shortcodes\latest_titles_shortcode');
+add_shortcode('aldine_page_section', [\Aldine\Shortcodes::class, 'pageSection']);
+add_shortcode('aldine_call_to_action', [\Aldine\Shortcodes::class, 'callToAction']);
+add_shortcode('latest-titles', [\Aldine\Shortcodes::class, 'latestTitlesShortcode']);
 
 // Catalog page: Network admin controls.
-add_action('admin_enqueue_scripts', '\Aldine\Admin\admin_scripts');
-add_action('wp_ajax_pressbooks_aldine_update_catalog', '\Aldine\Admin\update_catalog');
-add_filter('wpmu_blogs_columns', '\Aldine\Admin\catalog_columns');
-add_action('manage_blogs_custom_column', '\Aldine\Admin\catalog_column', 1, 3);
-add_action('manage_sites_custom_column', '\Aldine\Admin\catalog_column', 1, 3);
+add_action('admin_enqueue_scripts', [\Aldine\Admin\Catalog::class, 'adminScripts']);
+add_action('wp_ajax_pressbooks_aldine_update_catalog', [\Aldine\Admin\Catalog::class, 'updateCatalog']);
+add_filter('wpmu_blogs_columns', [\Aldine\Admin\Catalog::class, 'catalogColumns']);
+add_action('manage_blogs_custom_column', [\Aldine\Admin\Catalog::class, 'catalogColumn'], 1, 3);
+add_action('manage_sites_custom_column', [\Aldine\Admin\Catalog::class, 'catalogColumn'], 1, 3);
 
 // Remove unwanted menu pages.
-add_action('admin_menu', '\Aldine\Actions\remove_menu_items');
+add_action('admin_menu', [\Aldine\Actions::class, 'removeMenuItems']);
 
 // Remove unwanted actions.
 remove_action('before_delete_post', '_reset_front_page_settings_for_post');
