--- src/Tags/namespace.php
@@ -1,9 +1,0 @@
-<?php
-
-/**
- * Aldine Template Tags
- *
- * @package Aldine
- */
-
-namespace Aldine\Tags;
