+++ tests/test-filters.php
@@ -6,7 +6,7 @@
  * @package Pressbooks_Aldine
  */
 
-use function Aldine\Filters\register_query_vars;
+use Aldine\Filters;
 
 /**
  * Filters test case.
@@ -15,7 +15,7 @@ class FiltersTest extends WP_UnitTestCase
 {
     public function test_register_query_vars()
     {
-        $vars = register_query_vars([]);
+        $vars = Filters::registerQueryVars([]);
         $this->assertContains('license', $vars);
         $this->assertContains('subject', $vars);
     }
